+++ src/locations/locations.service.ts
@@ -1,10 +1,11 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException, ForbiddenException } from '@nestjs/common';
 import { CreateLocationInput } from './dto/create-location.input';
 import { UpdateLocationInput } from './dto/update-location.input';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Not, Repository } from 'typeorm';
+import { Repository } from 'typeorm';
 import { Location } from './entities/location.entity';
 import { DecodedToken } from 'src/auth/interfaces/auth.interface';
+import { Action, AppAbility } from 'src/casl/interfaces/casl.interface';
 
 @Injectable()
 export class LocationsService {
@@ -28,12 +29,27 @@ export class LocationsService {
     return this.locationsRepository.find({ where: { created_by: id } })
   }
 
-  findOne(id: number) {
-    return this.locationsRepository.findOne({ where: { id } })
+  async findOne(ability: AppAbility, id: number) {
+    const location = await this.locationsRepository.findOne({ where: { id } })
+    if (!location) {
+      throw new NotFoundException()
+    }
+
+    if (!ability.can(Action.Read, location)) {
+      throw new ForbiddenException()
+    }
+
+    return location;
   }
 
-  async update(id: number, updateLocationInput: UpdateLocationInput) {
+  async update(ability: AppAbility, id: number, updateLocationInput: UpdateLocationInput) {
     const { name, address } = updateLocationInput;
+    const location = await this.findOne(ability, id);
+
+    if (!ability.can(Action.Update, location)) {
+      throw new ForbiddenException();
+    }
+
     const result = await this.locationsRepository
       .createQueryBuilder().update(Location).set({
         name,
@@ -42,14 +58,12 @@ export class LocationsService {
       .where({ id })
       .returning(["name", "address", "id"]).execute()
 
-    if (!result.affected) {
-      throw new NotFoundException()
-    }
-
     return result.raw[0]
   }
 
-  remove(id: number) {
+  async remove(ability: AppAbility, id: number) {
+    const location = await this.findOne(ability, id);
+    if (!ability.can(Action.Delete, location)) throw new ForbiddenException();
     return this.locationsRepository.delete(id)
   }
 }
